feature: 1 to 1 のサンプルを作成
