+++ lib/components/components.dart
@@ -1,0 +1,2 @@
+export 'package:router_game_f/components/packet.dart';
+export 'package:router_game_f/components/pc_node.dart';
